% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNT.PV.wtd.mean.R
\name{CNT.PV.wtd.mean}
\alias{CNT.PV.wtd.mean}
\title{Weighted Mean of PISA Scores per Country Function}
\usage{
CNT.PV.wtd.mean(
  data,
  PVtype,
  weight = "W_FSTUWT",
  PVnumber = 10,
  CNTlabel = "CNT"
)
}
\arguments{
\item{data}{Here input the dataframe which contains Plausible Values and weights.}

\item{PVtype}{String which contains information on which Plausible Values you want to use (e.g. math, scie...).}

\item{weight}{String containing the name of the weight variable. Defaults to W_FSTUWT.}

\item{PVnumber}{Number which represents the number of Plausible Values. Defaults to 10.}

\item{CNTlabel}{String containg the name of the varaible with country labels/names. Defaults to CNT.}
}
\value{
Dataframe which contains a list of countries each with a weighted score and number of observations.
}
\description{
This function computes the weighted mean of PISA scores which are computed
based on all Plausible Values for each unique country and returns a dataframe with the list of
countries, weighted scores for each country and number of observations.
This function assumes that Plausible Values are labeled as PVXSUBJECT (e.g. PV1MATH, PV2SCIE...).
}
