% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.PV.wtd.mean.R
\name{Group.PV.wtd.mean}
\alias{Group.PV.wtd.mean}
\title{Weighted Mean of PISA Scores per Group Function}
\usage{
Group.PV.wtd.mean(
  data,
  PVtype,
  weight = "W_FSTUWT",
  PVnumber = 10,
  GroupLabel = "CNT"
)
}
\arguments{
\item{data}{Here input the data frame which contains Plausible Values and weights.}

\item{PVtype}{String which contains information on which Plausible Values you want to use (e.g. math, scie...).}

\item{weight}{String containing the name of the weight variable. Defaults to W_FSTUWT.}

\item{PVnumber}{Number which represents the number of Plausible Values. Defaults to 10.}

\item{GroupLabel}{String containing the name of the variable that will serve as a grouping variable (e.g., CNT, gender...). Defaults to CNT.}
}
\value{
Data frame which contains a list of countries each with a weighted score and number of observations.
}
\description{
This function computes the weighted mean of PISA scores which are computed
based on all Plausible Values for each unique country and returns a data frame with
weighted scores for each country and number of observations grouped by a variable.
This function assumes that Plausible Values are labeled as PVXSUBJECT (e.g. PV1MATH, PV2SCIE...).
}
