% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.wtd.mean.R
\name{Group.wtd.mean}
\alias{Group.wtd.mean}
\title{Weighted Means per Group Function}
\usage{
Group.wtd.mean(
  data,
  varMean,
  Grouplabel = "CNT",
  weight = "W_FSTUWT",
  showCI = TRUE,
  alpha = 0.95,
  repwt = "W_FSTR"
)
}
\arguments{
\item{data}{Here input the data frame which contains the variable for which you need the country means, final student weight, and replicate weights.}

\item{varMean}{String containing the name of the variable for which you need to compute the country means.}

\item{Grouplabel}{String containing the name of the variable with country labels/names. Defaults to CNT.}

\item{weight}{String containing the name of the final student weight variable. Defaults to W_FSTUWT.}

\item{showCI}{Boolean. If true, confidence intervals are shown in the return. Defaults to TRUE.}

\item{alpha}{Decimal number affecting the confidence interval levels. Defaults to 0.95 (95\% confidence intervals).}

\item{repwt}{String containing the prefix of the replicate weights variables. Defaults to W_FSTR.}
}
\value{
Data frame which contains a list of countries each with a weighted mean, weighted standard error, number of observations, and optionally the confidence intervals.
}
\description{
This function computes the weighted means of a variable for all unique
countries in the data frame and returns a data frame with the
weighted means, weighted standard errors,
number of observations, and optionally also provides confidence intervals grouped by a variable.
The function assumes that the variable varMean does not contain NA codes.
}
