% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overclaiming.calc.R
\name{overclaiming.calc}
\alias{overclaiming.calc}
\title{Overclaiming Calculations Function}
\usage{
overclaiming.calc(data, existIt, nonexistIt, scaleStart = 1, scaleStop = 5)
}
\arguments{
\item{data}{Here input the data frame which contains all overclaiming items.}

\item{existIt}{Vector containing strings with variable names for existing items.}

\item{nonexistIt}{Vector containing strings with variable names for nonexistent items.}

\item{scaleStart}{Number which represents the code for the start of the answering scale (e.g., "Never heard" coded as 1). Defaults to 1.}

\item{scaleStop}{Number which represents the code for the end of the answering scale (e.g., "Know it well" coded as 5). Defaults to 5.}
}
\value{
Returns appended imputed data frame which now contains variables needed for computation of overclaiming indices.
}
\description{
This function takes a data frame and computes necessary calculation for overclaiming analysis.
After calculation it returns an appended data frame with overclaiming variables.
It creates dummy variables, variables for computation of overclaiming indices
Proportion of Hits (PH), Proportion of False Alarms (PFA), Index of Accuracy (IA),
and Index of Exaggeration (IE).
This function does not assume absence of NA codes. It, however, assumes that
the answer codes are in a series (e.g., 1, 2, 3, 4, 5). It also assumes that
1 represents "never heard" answer and 5 being the "most familiar" answer. Further,
it assumes that respondents without answers to any OCT item are excluded.
}
