% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISA.wt.multilvl.transform.R
\name{PISA.wt.multilvl.transform}
\alias{PISA.wt.multilvl.transform}
\title{PISA weight multilevel transform function}
\usage{
PISA.wt.multilvl.transform(data, weight = "W_FSTUWT", CNTlabel = "CNT")
}
\arguments{
\item{data}{Here input the data frame which contains the weights.}

\item{weight}{String containing the name of the weight variable. Defaults to W_FSTUWT.}

\item{CNTlabel}{String containing the name of the variable with country labels/names. Defaults to CNT.}
}
\value{
Returns data frame with a "multilvl_wt" column where transformed weights are located
}
\description{
This function takes the Final student weight and transforms it to be usable
in multilevel analyses. The final student weights are transformed in a way
that the sum of weights per each country is constant and equal to the average
number of respondents per country.
}
