% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISA.wtd.std.error.R
\name{PISA.wtd.std.error}
\alias{PISA.wtd.std.error}
\title{PISA Weighted Standard Error Function}
\usage{
PISA.wtd.std.error(data, varS, finalwt = "W_FSTUWT", repwt = "W_FSTR")
}
\arguments{
\item{data}{Here input the data frame which contains the variable for which you need standard error, final student weight, and replicate weights.}

\item{varS}{String containing the name of the variable for which you need to compute the standard error.}

\item{finalwt}{String containing the name of the final student weight variable. Defaults to W_FSTUWT.}

\item{repwt}{String containing the prefix of the replicate weights variables. Defaults to W_FSTR.}
}
\value{
List containing the standard error (std_error) and sampling variance (sampVar).
}
\description{
This function computes the weighted standard error for PISA variables. It
uses the final student weight and replicate weights for computation according
to PISA data analysis manual.
The principle of the function is that final mean is computed using the final
student weight and then 80 means (weighted by replicate weights) are computed.
Then the sampling variance and standard error are computed.
The function assumes that the variable varS does not contain NA codes.
}
