% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PV.correlation.R
\name{PV.correlation}
\alias{PV.correlation}
\title{Correlation with PISA Score Function}
\usage{
PV.correlation(data, cor1, PVtype, PVnumber = 10)
}
\arguments{
\item{data}{Here input the data frame which contains the variable for which you need the correlation and Plausible Values.}

\item{cor1}{String which contains the variable name for which you want the correlation value.}

\item{PVtype}{String which contains information on which Plausible Values you want to use (e.g. math, scie...).}

\item{PVnumber}{Number which represents the number of Plausible Values. Defaults to 10.}
}
\value{
List which contains the correlation value, and confidence intervals.
}
\description{
This function computes the (unweighted) correlation of a variable with PISA scores
which are computed based on all Plausible Values.
This function assumes that Plausible Values are labeled as PVXSUBJECT (e.g. PV1MATH, PV2SCIE...).
It also assumes that variable does not contain NA codes.
}
