% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PV.wtd.mean.R
\name{PV.wtd.mean}
\alias{PV.wtd.mean}
\title{Weighted Mean of PISA Scores Function}
\usage{
PV.wtd.mean(data, PVtype, weight = "W_FSTUWT", PVnumber = 10)
}
\arguments{
\item{data}{Here input the dataframe which contains Plausible Values and weights.}

\item{PVtype}{String which contains information on which Plausible Values you want to use (e.g. math, scie...).}

\item{weight}{String containing the name of the final student weight variable. Defaults to W_FSTUWT.}

\item{PVnumber}{Number which represents the number of Plausible Values. Defaults to 10.}
}
\value{
Returns weighted mean score.
}
\description{
This function computes the weighted mean of PISA scores which are computed
based on all Plausible Values.
This function assumes that Plausible Values are labeled as PVXSUBJECT (e.g. PV1MATH, PV2SCIE...).
}
