% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicaWTlistfc.R
\name{replicaWTlistfc}
\alias{replicaWTlistfc}
\title{Replicate Weight List Function}
\usage{
replicaWTlistfc(repwt = "W_FSTR")
}
\arguments{
\item{repwt}{String containing the prefix of the replicate weights variables. Defaults to W_FSTR.}
}
\value{
List with 80 replicate weight labels.
}
\description{
This is an auxiliary function that creates a list of 80 replicate weights
based on the label input.
}
