% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT.moving.windows.R
\name{RT.moving.windows}
\alias{RT.moving.windows}
\title{Response Time Moving Windows Function}
\usage{
RT.moving.windows(
  data,
  timeVarLabel,
  funcName,
  timeStop,
  timeStart = 0,
  intLength = 2000,
  windowShift = 100,
  endInclusive = TRUE,
  ...
)
}
\arguments{
\item{data}{Here input the data frame which contains needed variables.}

\item{timeVarLabel}{String. The name of the time variable.}

\item{funcName}{String. Here you enter the name of the function you want to apply on the examined variable in each time window. Include required variables in calling of this function. If data frame is required, the function will automatically pass the data variable.}

\item{timeStop}{Number. Indicates the end time of your analysis.}

\item{timeStart}{Number. Indicates the starting time of your analysis. Defaults to 0.}

\item{intLength}{Number. Indicates the length of the time window. E.g. If your time variable is in milliseconds then value 2000 will mean 2s intervals (e.g. 0-2, 0.1-2.1, etc.). Defaults to 2000.}

\item{windowShift}{Number. Indicates by how much the window will shift. E.g. If your time variable is in milliseconds then value 100 will mean shift by 0.1s (e.g., 0-2, 0.1-2.1, etc.). Defaults to 100.}

\item{endInclusive}{Boolean. If TRUE then the time intervals will be inclusive of the end of each interval window (e.g. 0-2s inclusive of 2s but not of 0s). If FALSE, then it is inclusive of the beginning (e.g. 0-2s, inclusive of 0s but not of 2s). Defaults to TRUE.}

\item{...}{Add all required arguments for a function in funcName.}
}
\value{
Returns data frame where rows are each time interval window with two (or more) columns - timeInt which shows individual time windows, and funcName which provides the result of the function in funcName applied on data in the examined time window. If the applied function returns more than one parameter, the data frame is expanded to present all parameters.
}
\description{
This function performs the response time analysis using the moving window
method. This method analyzes response time windows of a set length and moves
the windows by a specified jump. This creates overlapping intervals which
allows to see how an observed variable develops over time. The smaller the
window, the more detailed observation of the development of the variable.
This function performs the analysis in individual windows and provides you
with a data frame with results.
}
