% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tval.pval.conv.R
\name{tval.pval.conv}
\alias{tval.pval.conv}
\title{T-value to P-value Conversion Function}
\usage{
tval.pval.conv(t_value, df = 1e+05, tail = "two.sided")
}
\arguments{
\item{t_value}{Number that represents a t-value.}

\item{df}{Number that represents degrees of freedom. Defaults to 100000.}

\item{tail}{String that represents the hypothesis. Either two.sided, greater or less. Defaults to two.sided.}
}
\value{
Returns the p-value.
}
\description{
This function takes a t-value and converts it to a p-value.
}
