% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tval.Stars.R
\name{tval.Stars}
\alias{tval.Stars}
\title{T-value Stars Function}
\usage{
tval.Stars(value, tval, decplaces = 3)
}
\arguments{
\item{value}{Number containing the value (e.g. estimate).}

\item{tval}{Number containg the t-value associated with value number.}

\item{decplaces}{Number which represents the rounding of numbers to a specific decimal place. Defaults to 3.}
}
\value{
String with the value input and stars assigned pased on significance level, rounded to the specified number of decimal places.
}
\description{
This funtion takes in a number and its corresponding t-value.
Returns a number with attached stars based on significance level.
p < 0.001 ***, p < 0.01 **, p < 0.05 *, p < 0.1 . or without stars if the p >= 0.1.
It is used, for example, when you run a model which gives you estimate together
with its t-value without star symbols that represent the significance level.
It is an auxiliary function for easy making of APA style tables.
}
