% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISA.wt.multilvl.transform.R
\name{PISA.wt.multilvl.transform}
\alias{PISA.wt.multilvl.transform}
\title{PISA weight multilevel transform function}
\usage{
PISA.wt.multilvl.transform(data, weight = "W_FSTUWT", CNTlabel = "CNT")
}
\arguments{
\item{data}{Here input the data frame which contains the weights.}

\item{weight}{String containing the name of the weight variable. Defaults to W_FSTUWT.}

\item{CNTlabel}{String containing the name of the variable with country labels/names. Defaults to CNT.}
}
\value{
Returns the original data frame with a "multilvl_wt" column where transformed weights are located. It also provides "std_wt" column which is an intermediate step for calculation of "multilvl_wt". These weights are a transformation of the final student weight in a way that the total sum of all weights equals the number of respondents in the whole data frame (i.e. the sum of weights for each country is the number of respondents in that country).
}
\description{
This function takes the final student weight and transforms it to be usable
in multilevel analyses as per the PISA data analysis manual. The final student
weights are transformed in a way that the sum of weights per each country is
constant and equal to the average number of respondents per country ("multilvl_wt"). It also
computes transformation of the final student weight where the total sum of
weights is equal to the number of respondents in the data frame (i.e., the
sum of weights for each country equals the number of respondents in that country)
}
